+++ src/node/blockstorage.cpp
@@ -17,6 +17,7 @@
 #include <signet.h>
 #include <streams.h>
 #include <undo.h>
+#include <util/batchpriority.h>
 #include <util/fs.h>
 #include <util/syscall_sandbox.h>
 #include <util/system.h>
