+++ src/util/system.cpp
@@ -7,16 +7,9 @@
 
 #include <logging.h>
 #include <util/string.h>
-#include <util/syserror.h>
 #include <util/time.h>
 
-#if (defined(__FreeBSD__) || defined(__OpenBSD__) || defined(__DragonFly__))
-#include <pthread.h>
-#include <pthread_np.h>
-#endif
-
 #ifndef WIN32
-#include <sched.h>
 #include <sys/stat.h>
 #else
 #include <codecvt>
@@ -111,15 +104,4 @@ int GetNumCores()
 int64_t GetStartupTime()
 {
     return nStartupTime;
-}
-
-void ScheduleBatchPriority()
-{
-#ifdef SCHED_BATCH
-    const static sched_param param{};
-    const int rc = pthread_setschedparam(pthread_self(), SCHED_BATCH, &param);
-    if (rc != 0) {
-        LogPrintf("Failed to pthread_setschedparam: %s\n", SysErrorString(rc));
-    }
-#endif
 }
